+++ bio/multiqc/wrapper.py
@@ -11,17 +11,26 @@ from os import path
 from snakemake.shell import shell
 
 
-input_dirs = set(path.dirname(fp) for fp in snakemake.input)
+extra = snakemake.params.get("extra", "")
+# Set this to False if multiqc should use the actual input directly
+# instead of parsing the folders where the provided files are located
+use_input_files_only = snakemake.params.get("use_input_files_only", False)
+
+if not use_input_files_only:
+    input_data = set(path.dirname(fp) for fp in snakemake.input)
+else:
+    input_data = set(snakemake.input)
+
 output_dir = path.dirname(snakemake.output[0])
 output_name = path.basename(snakemake.output[0])
 log = snakemake.log_fmt_shell(stdout=True, stderr=True)
 
 shell(
     "multiqc"
-    " {snakemake.params}"
+    " {extra}"
     " --force"
     " -o {output_dir}"
     " -n {output_name}"
-    " {input_dirs}"
+    " {input_data}"
     " {log}"
 )